requirements.txt
psycopg2
django
sqlalchemy
gino
asyncpg
aiogram
environs
django-import-export
django-jazzmin